Extended Landau Gauge
We extended the way the constant part of the gauge field is calculated
so that we can have the extended Landau Gauge and also, more
specifically, with the aim of having a rotationally symmetric gauge.
This came at the cost of some performance since we now need all lattice
sites to have an additional y coordinate, and also since we have to
calculate a bit more when finding gauge field for the free energy. We
have shown that this gauge will not affect the Maxwell term so this is
still calculated only considering the fluctuating part of the gauge
field. This helped performance slightly.