Extending next nearest neighbors
We added r+x+y and r-x-y to the next nearest neighbors of each lattice
site. This neccessitated the addition to two additional shell edges as
well as change in the update protocol such that these edges where
updated individually in case the energy depended on these new next
nearest neighbors.